Added the basics.
Added the necessary files, such as an html file for the popup, a css file for the styling, a javascript file to display the html popup, and a manifest file. Wrote basic code for each file to display a static image and basic text.  Implementation of pulling a random meme from the web is next.